requirements.txt
python-telegram-bot==13.15
tweepy==4.14.0
python-dotenv==1.0.1
requests==2.31.0
